+++ src/routes/Overview.svelte
@@ -6,6 +6,7 @@
   import MonthSummaryCard from './overview/MonthSummaryCard.svelte';
   import SalaryCard from './overview/SalaryCard.svelte';
   import InvestedCard from './overview/InvestedCard.svelte';
+  import DebtCard from './overview/DebtCard.svelte';
   import { appStore, settings } from '@/lib/appStore';
   import { monthKey as toMonthKey } from '@/lib/db/months';
   import { monthActivity } from '@/lib/activity';
@@ -26,6 +27,8 @@
     <SalaryCard />
     <InvestedCard />
   </div>
+
+  <DebtCard />
 
   <Card title={$t('overview.activity')}>
     <ActivityList entries={recent} language={$settings.language} />